requirements.txt
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
billiard==3.3.0.23
celery==3.1.24
decorator==4.0.10
Django==1.10
django-activity-stream==0.6.3
django-allauth==0.26.1
django-anymail==0.6.1
django-contrib-comments==1.7.2
django-filter==0.14.0
django-jsonfield==1.0.1
django-liked==1.0.2
djangorestframework==3.4.6
imageio==1.6
ipython==5.0.0
ipython-genutils==0.1.0
kombu==3.0.37
Markdown==2.6.6
moviepy==0.2.2.11
numpy==1.11.2
oauthlib==1.1.2
pathlib2==2.1.0
pexpect==4.2.0
pickleshare==0.7.3
Pillow==3.3.1
prompt-toolkit==1.0.3
ptyprocess==0.5.1
Pygments==2.1.3
python-openid==2.2.5
pytz==2016.6.1
requests==2.11.0
requests-oauthlib==0.6.2
simplegeneric==0.8.1
six==1.10.0
tqdm==4.8.4
traitlets==4.2.2
wcwidth==0.1.7
